# change_sub OPTIMIZE write event only if needed