@irfanHaslanded 2025-05-16
### **Test Summary**
With these changes, the number of futex syscalls seems to reduce by nearly 30%, and the number of unnecessary wakes (or errors) is reduced significantly.

---

Running only the `test_many_oper_change_subs` under strace, with below diff
```diff
$ git diff tests/
diff --git a/tests/perf.c b/tests/perf.c
index bc698d4b..142455ce 100644
--- a/tests/perf.c
+++ b/tests/perf.c
@@ -987,6 +987,8 @@ sysrepo_destroy(void)
 }

 struct test tests[] = {
+    {"many oper change_subs", setup_empty_oper, test_many_oper_change_subs, teardown_empty},
+#if 0
     {"get tree", setup_running, test_get_tree, teardown_running},
     {"get item", setup_running, test_get_item, teardown_running},
     {"get tree hash", setup_running, test_get_tree_hash, teardown_running},
@@ -1009,6 +1011,7 @@ struct test tests[] = {
     {"modify an item cached", setup_running_cached, test_item_modify, teardown_running},
     {"remove an item", setup_running, test_item_remove, teardown_running},
     {"remove an item cached", setup_running_cached, test_item_remove, teardown_running},
+#endif
 };
```
---
## with these changes 0.467227 s (60% of system calls time spent in futex)

```sh
$ make -j sr_clean test_clean all > /dev/null 2>&1 && strace -c  ./tests/sr_perf 100 1 2>&1 | grep -v '0.00'

| Options

  Data set size      : 100
  Each test executed : 1 time


| Performance tests

  JSON DS file
 ___________________________________________________________________
|                                                                   |
| test name                          |         time |    comparison |
|___________________________________________________________________|
|                                                                   |
| many oper change_subs ............ |   0.467227 s |       1.000 x |
|___________________________________________________________________|


All comparisons refer to how many times faster (green) or slower (red) the current plugin is compared to the first plugin.

 % time     seconds  usecs/call     calls    errors syscall
------ ----------- ----------- --------- --------- ----------------
 59.86    0.066825          55      1194         4 futex
------ ----------- ----------- --------- --------- ----------------

```
---
## without these changes 0.606427 s (84% of system calls time spent in futex)

```sh
$ make -j sr_clean test_clean all > /dev/null 2>&1 && strace -c  ./tests/sr_perf 100 1 2>&1 | grep -v '0.00'

| Options

  Data set size      : 100
  Each test executed : 1 time


| Performance tests

  JSON DS file
 ___________________________________________________________________
|                                                                   |
| test name                          |         time |    comparison |
|___________________________________________________________________|
|                                                                   |
| many oper change_subs ............ |   0.606427 s |       1.000 x |
|___________________________________________________________________|


All comparisons refer to how many times faster (green) or slower (red) the current plugin is compared to the first plugin.

 % time     seconds  usecs/call     calls    errors syscall
------ ----------- ----------- --------- --------- ----------------
 84.16    0.195312         107      1825       205 futex
------ ----------- ----------- --------- --------- ----------------
```